+++ src/net/sf/mpxj/primavera/PrimaveraPMFileWriter.java
@@ -746,17 +746,21 @@ public final class PrimaveraPMFileWriter extends AbstractProjectWriter
          FieldType fieldType = cf.getFieldType();
          if (fieldType != null && type == fieldType.getFieldTypeClass())
          {
-            UDFAssignmentType udf = m_factory.createUDFAssignmentType();
-            udf.setTypeObjectId(FieldTypeHelper.getFieldID(fieldType));
-            setUserFieldValue(udf, fieldType.getDataType(), mpxj.getCachedValue(fieldType));
-            out.add(udf);
+            Object value = mpxj.getCachedValue(fieldType);
+            if (FieldTypeHelper.valueIsNotDefault(fieldType, value))
+            {
+               UDFAssignmentType udf = m_factory.createUDFAssignmentType();
+               udf.setTypeObjectId(FieldTypeHelper.getFieldID(fieldType));
+               setUserFieldValue(udf, fieldType.getDataType(), value);
+               out.add(udf);
+            }
          }
       }
       return out;
    }
 
    /**
-    * Sets the vaue of a UDF.
+    * Sets the value of a UDF.
     *
     * @param udf user defined field
     * @param dataType MPXJ data type
@@ -767,42 +771,66 @@ public final class PrimaveraPMFileWriter extends AbstractProjectWriter
       switch (dataType)
       {
          case DURATION:
-            udf.setTextValue(value != null ? ((Duration) value).toString() : "");
+         {
+            udf.setTextValue(((Duration) value).toString());
             break;
+         }
+
          case CURRENCY:
-            if (value != null && !(value instanceof Double))
+         {
+            if (!(value instanceof Double))
             {
                value = Double.valueOf(((Number) value).doubleValue());
             }
             udf.setCostValue((Double) value);
             break;
+         }
+
          case BINARY:
+         {
             udf.setTextValue("");
             break;
+         }
+
          case STRING:
-            // write empty string instead of null to make sure we get a value
-            // node in the XML output
-            udf.setTextValue(value != null ? (String) value : "");
+         {
+            udf.setTextValue((String) value);
             break;
+         }
+
          case DATE:
+         {
             udf.setStartDateValue((Date) value);
             break;
+         }
+
          case NUMERIC:
-            if (value != null && !(value instanceof Double))
+         {
+            if (!(value instanceof Double))
             {
                value = Double.valueOf(((Number) value).doubleValue());
             }
             udf.setDoubleValue((Double) value);
             break;
+         }
+
          case BOOLEAN:
+         {
             udf.setIntegerValue(BooleanHelper.getBoolean((Boolean) value) ? Integer.valueOf(1) : Integer.valueOf(0));
             break;
+         }
+
          case INTEGER:
          case SHORT:
+         {
             udf.setIntegerValue(NumberHelper.getInteger((Number) value));
             break;
+         }
+
          default:
+         {
             throw new RuntimeException("Unconvertible data type: " + dataType);
+         }
       }
    }